+++ addon/components/object-list-view-row.js
@@ -69,8 +69,13 @@ export default FlexberryBaseComponent.extend({
   */
   _records: Ember.computed(() => Ember.A()),
 
-  rendered: false,
+  /**
+    Flag used to start render row content.
 
+    @property doRenderData
+    @type Boolean
+    @default false
+  */
   doRenderData: false,
 
   /**
@@ -113,6 +118,7 @@ export default FlexberryBaseComponent.extend({
             key: Ember.guidFor(record),
             data: record,
             config: config,
+            doRenderData: true
           });
 
           this.get('_records').pushObject(newRecord);
@@ -217,11 +223,4 @@ export default FlexberryBaseComponent.extend({
   didInsertElement() {
     Ember.$('.object-list-view-menu > .ui.dropdown').dropdown();
   },
-
-  didRender() {
-    this._super(...arguments);
-    if (this.doRenderData) {
-      this.set('rendered', true);
-    }
-  }
 });